Remove unused color depth option
Since the migration to SDL2 FreeOrion never exposed the option to set
the color depth.  Still nobody complained about the not working option.

Also a lot of post 2003 commercial games don't expose a setting for this
but rather let the OS autoselect the best color depth for the current
display.