Move FO CMake test dir inclusion in front of GG inclusion
When including the GG sources first the GG `unittest` target is
generated within a subdirectory.  This removes the ability to remove the
FreeOrion unit test to either append themselves to the existing
`unittest` target (because it is located in a subdirectory) or to create
their own `unittest` target (because that would cause a name clash).