Use emplace_back when placing std::tuple into std::vector
Older GCC versions are not able to convert a initializer list to
a tuple when using push_back but rather try to pass the initializer
list as the first parameter of a tuple construction.  This leads to
missmatching types of the std::vector::value_type and the passed
push_back parameter.

Fixes #1357