Use std::create_tuple when returning a std::tuple
Older GCC versions are not able to convert a initializer list to
a tuple when just returning an initializer list. They rather try to
pass the initializer list as the first parameter of a tuple
construction.  This leads to missmatching types between the return
statement and the function return declaration.

Fixes #1357