Fix crashes in ModalListPicker in the multiplayer lobby.
If the parent of a ModalListPicker that is dropped down and in a modal
event loop is destroyed, then signaling or returning from the Run() and
accessing any members of the parent results in a segfault.

It is not possible to throw an exception and unroll the modal event pump
cleanly.

This commit uses a pair of shared_ptrs to the ModalListPicker instance
to determine if the parent has been destroyed before signaling or
returning from Run().