Fix FleetWnd doesn't update StatsIcon went ship with fighters merged in.
The problem is that when the FleetDataPanel is initialized it determines
which StatisticIcons are included based on its fleet's current
capabilities.  When a ship is transferred in with new capabilities, like
fighters, it does not have a StatisticIcon to update with the new
values.

This commit unconditionally, includes all StatisticIcons possible for
fleets and then conditionally displays those with non null values.