Make MessageQueue::PopFront return optional<Message>.
Previously, ClientNetworking would check if the queue was empty and then
pop the first item under two different mutexes.  This meant that it was
possible for the queue state to change between check and pop.

Returning optional<Message> fixes that bug.  The check and pop occur under
the same mutex.

If this is not correct, please remove the file # .git/CHERRY_PICK_HEAD #
and try again.