Make MessageQueue::GetFirstSynchronousMessage not hang.
MessageQueue::GetFirstSynchronousMessage would hang if the client
networking shutdown the receive queue, while a synchronous message was
pending.

This commit uses the m_have_synchronous condition from StopGrowing() to
signal the waiting GetFirstSynchronousMessage function, which returns
boost::none to indicate that there is no synchronous response to the
given message.