Workaround static de-initialization crashes in boost::log.
There are two problems with static de-initialization in boost::log.

The first is that boost::log uses boost::filesystem to log to files.
boost::filesystem uses a static std::locale object.  There is no way to
influence the order of static de-initialization. Intermittently during
static de-initialization boost::log crashes while flushing the file.

This bug is described in:
http://www.boost.org/doc/libs/1_64_0/libs/log/doc/html/log/rationale/why_crash_on_term.html
https://svn.boost.org/trac/boost/ticket/8642
https://svn.boost.org/trac/boost/ticket/9119

The second bug was reported in PR #1481.

During static de-initialization boost::log exhausts the thread local
storage (TLS) and crashes.

The solution in this commit, is to close the log file before the start
of static de-initialization at the end of main().  Any remaining logs
will be logged to the console.