Fix universe XTypeManager singleton enforcement of single constructor use.
The problem is caused by 4 facts.

1. C++ re-executes the constructor of any static function variable that
   has failed every time that execution passes over the static
   variable declaration.

2. The singletons, BuildingTypeManager, FieldTypeManager,
   PartTypeManager, HullTypeManager, PredefinedShipDesignManager and
   TechManager all enforce the single use of their constructor by
   storing the address of the first "this" pointer and throwing a
   runtime_error if an additional use of the constructor occurs with a
   different address.

3. The constructors also throw other errors on failure.

4. The constructors capture the "this" pointer before the other code
   that throws.

Therefore on every subsequent execution of these constructors they throw
an error indicating that there are multiple singletons, when there is
not even one.

The fix is to move the capture of the "this" pointer to the end of the
constructor, after any code that could cause failure of the constructor.