Refactor some Universe XTypeManager parsers to use the ParseImpl functors.
The parsers for BuildingType, FieldType, PartType, HullType and
ShipDesign, use custom functors which check if the type Name() was not
unique and throw an exception or insert the type into a map.

This is the only code in the parse namespace that uses exceptions to
indicate errors.

These parsers are used by the singletons, BuildingTypeManager,
FieldTypeManager, PartTypeManager, HullTypeManager and
PredefinedShipDesignManager.  The singletons do not have a predictable
point of construction, particularly if they have or will throw an
exception, which may occur inside the python code where it is difficult
to handle correctly.

This commit refactors those custom functors so that errors are reported
in the same way as the boost::spirit error handler, with the Logger.

The custom functor is replaced with two generic functors.

The first, 'is_unique', checks if the name is not unique and causes the
parse of the name to fail.

The second, 'insert', does the insert assuming that the check already
completed successfully.

The result is that all parse errors are indicated to the calling
function in the same manner, through the log.