Refactor TechParser parsers to use the ParseImpl functors.
The parsers for Tech, uses custom functors which check if the type
Name() and category was not unique and throws an exception or inserts
the type into a map.

This is the only code in the parse namespace that uses exceptions to
indicate errors.

The parsers are used by the singleton, TechManager.  It does not have a
predictable point of construction, particularly if throwing exceptions,
which may occur inside the python code where it is difficult to handle
correctly.

This commit refactors those custom functors so that errors are reported
in the same way as the boost::spirit error handler, with the Logger.

The custom functors are replaced by either the generic functor,
'is_unique' or 3 new functors.

'check_tech', checks if a tech name is unique and causes the whole tech
parse to fail otherwise.

'insert_tech' inserts the tech assuming that 'check_tech' has passed.

'is_unique', is used to check if the category name is unique.

'insert_category', inserts the category assuming that it is unique.
The result is that all parse errors are indicated to the calling
function in the same manner, through the log.