Make default logging threshold for Release be debug.
As discussed in issue #1575, there are currently not enough logs at the
info, warning and error level for the log files to be useful to testers.
Since we do not have separate targets for players, testers and devs,
this commit is setting all build types default logger thresholds to
debug.