In Logger.cpp fix late configuration of file sink front end.
The problem is that when a logger is first configured if the file sink
backend has not been created, then the logger was not being stored for
the late configuration when the file sink backend is created.

This meant that if a logger was created and produced its first log
during static initialization, then it would correctly send logs to the
console, but would not send logs to the log file when the log file was
configured.

This commit stores the logger for late configuration.