Add IDAllocator.<h/cpp>
IDAllocator will be used by the Universe to allocate ids for universe
objects and ship designs.

It partitions the address space in modulo (number of empires + 1)
sections to prevent any id contention between the clients and the
server.

This can fix the problems of:
a) synchronous polling for ids which is causing delay
b) and the assignment of the id and insertion into the object map not
being atomic.