Move Phototrophic effects evaluate after all other effects
Because the phototrophic effect activates only with a target population
greater than zero, it depends on the order of effect evaluations that
modify the target population.

Prior to this PR, there were effects that modify the target population
with the same priority as the phototrophic trait which is ill-defined
behavior.