Refactor python logging to use the same formatting for print and logger.
This commit refactors python/LoggingWrapper.cpp to format all python
loggers with the same "logger" : "Filename":"LineNo" - "msg" format as
other FreeOrion logs.

It changes the stdout and stderr (print) stream from the python code to
collect similar information (filename and line number) as the standard
library logger and send it to the C++ process.

It changes the logger handler to use the new C++ interface.