Add Universe::InsertNew<T>(Args...) and InsertTemp<T>(Args...)
InsertNew<T>() and InsertTemp<T>() construct and insert a universe
object with either a new id, or the temporary id.
This provides better control by the Universe of its own state.  This
prevents sending the id to the client and then the client
sending the (hopefully) same id back again.

This removes the need for separate generator functions for every variant
of a UniverseObject constructor.