Remove overly broad catch in HumanClientApp message queue.
The catch around message queue GetMessage was catching CleanQuit and
preventing the app from exiting if the exit signal arrived when the
message queue was being polled.

It is unclear what the original catch all was meant to catch.

This fixes a race condition between quitting the app and polling the
message queue that could prevent the app from exiting.