When server's python server fails, wait for client connect, then exit.
Shutting down immediately when the python server fails to initialize,
either leaves some AI processes hung, or kills AI processes in the midst
of their shutdown procedure.

Waiting for the clients to completely connect and then disconnecting the
is a more regular shutdown procedure. This procedure does not leave hung
AI processes and all AIs either exit due to being shutdown by the server
or because they could not connect to the server.