In configure_logging.py _stdXLikeStream dont attach context to stackless calls
_stdXLikeStream.write() includes filename, function name and line number
information taken from inspect.stack() in the information sent to the
C++ process.  PyErr_Print() is called directly by the C++ process
without a stack.  _stdXLikeStream.write() tries to access the stack that
does not exist and is unable to generate infomation for the log.

This commit changes _stdXLikeStream.write() to detect the empty stack
and pass empty strings for filename, function name and line number.

This fixes issue #1659, insufficient exception logging in AI scripts.