Change range based for loop variables from auto& to const auto&.
auto& can't bind to rvalues because the C++ spec requires it to not
extend the lifetime of the temporary, which immediately destroys itself.

Conversely, const auto& is required to extend the lifetime of the rvalue
until the end of the references scope.

This fixes the error, attempt to bind non-const lvalue to temporary
(rvalue) in both clang and gcc compilers.