Prevent double tag wrapping of CUILinkTextMultiEdit() links. (#1713)
CUILinkTextMultiEdit uses the underlying TextControl to store the text
processed to include the tags and overrides virtual SetText() to assign
the processed text.  The problem is that the underlying TextControl
calls SetText() internally when adjusting the scroll bars in its
CompleteConstruction().

This commit brackets CUIMultiEdit::CompleteConstruction() in
CUILinkTextMultiEdit::CompleteConstruction() in a
m_already_setting_text_so_dont_link guard to prevent the double
wrapping.