Clarify SaveFileDialog constructor(s) using enums and single a constructor.
The current SaveFileDialog constructors choose between saving and
loading, single player or multiplayer extension and server queries or
not with multiple constructors, a bool and a string extension that is
not always used.

This commit reduces this to a single constructor and enums for {Save,
Load} and {SinglePlayer, MultiPlayer} which is clearer.