workaround for rare AI bug from zombie AI empire
if an AI empire lost all its planets, but had a surviving military fleet
which then completed its mission without being destroyed, an error would
prevent the AI from properly reassigning the fleets and would get
reported to the player in the chat message window.

The error was in a subsection directed towards military backup for
exploration tasks.  This workaround essentially means that if the AI has
lost all of its planets then it will not provide military backup
for exploration efforts (which is probably in nearly all cases the best
decision).