Create ClientUI::GetFilenameWithSaveFileDialog().
This moves responsibility for the SaveFileDialog popup UI element out of
HumanClientApp and InGameMenu into a single location in ClientUI.  This
is consistent with other popup elements like the MessageWnd.

This will allow asynchronous loading of the save previews from the
server because the dialog will be persistent and accessible.