Fix AI colonization score for "negative" planet supply
In edge cases, the supply calculation yields a negative supply score for
a planet. However, planets can not negatively contribute to supply.

This bug happens often in the early game when an AI encounters a huge
planet within supply that it will not colonize as it incorrectly gives
the planet a negative supply score.