Remove unused synchronous messaging from client networking.
Synchronous messaging is broken in ways that can not be fixed.

If the client makes a synchronous request and the server FSM is in a
state that will not respond, the client hangs.

The implementation assumes that the client is single threaded.  If the
client is not single threaded and more than one thread makes a
synchronous request and the wrong thread receives any server response,
the client hangs.

If the server attempts to change the client FSM to a new state (like
shutdown) after a synchronous request is made, the client will ignore
the request.  Even if the server satisfies the synchronous request and
then sends the state change message, the client may send another
synchronous message before it handles the state change message.  Both
client and server hang.

The solution is to use TCP as an asynchronous protocol.  It works well.