Refactor most parsers to return content vs an empty container parameter.
This commit changes to

auto parsed_stuff = parse::stuff();

from

std::container parsed_stuff;
/* ignored error condition */ parse:stuff(parsed_stuff);

The idiom is clearer. The returned parsed content is returned.  The
container is constructed where it is used.

The ignored error code is a parser internal code indicating that every
parser stage succeeded, which the larger application does not care
about.

It will make code to parse the universe content on a different thread
and multiple times in the save executable clearer.