Start removing lexer singleton.
The lexer is a singleton.

The lexer is bound to the file being parsed and maintains the state of
the tokens being parsed.

Because the lexer is a singleton only one file can be parsed in any
executable at one time.  This means that there can be no recursive
parsing (macros that are not pre-parsed) nor any multi-threaded parsing.

This commit begins removing the lexer singleton by making its
constructor public.