Remove ill-performing "dithering" of foci
While the original idea may have been good and correct, nowadays the
code logic/math is simply wrong and we trade 1 RP for 5 PP or so.
Besides, temporarily dithering the focus means that other planets are
indirectly pushed to change focus as well because the ratios change.
This means even more focus change penalty.