Use more recent MacOSX image
Using the `xcode6.4` image on TravisCI pins the project to CMake 3.2.
Installing a more recent version of CMake via Homebrew will not use a
binary installation (Homebrew bottle) but instead a source installation.
This will prolong the build time and cause the build to fail by timeout.
Using a more recent MacOSX image will allow to use a binary build of
CMake and cutting down the build time.