Force ‘Release’ build config for Xcode generator
When using the Xcode generator, including ‘Debug’ into CMAKE_CONFIGURATION_TYPES results in the Debug build configuration always used when building the generated Xcode project via CMD-B, regardless what CMAKE_BUILD_TYPE is set to. The only way to prevent this and force Xcode to actually use the Release build configuration is to set CMAKE_CONFIGURATION_TYPES to ‘Release’ only.

As the resulting app of a debug build on OSX is more than twice the size as that of a non-debug build, also significantly slower, and you won’t need debug builds with Xcode most of the time, set CMAKE_CONFIGURATION_TYPES to ‘Release’ only when the Xcode has been specified.