Fix bugs in TruncateRouteToEndAtSystem
This changes the function name from Fleet::ShortenRouteToEndAtSystem()
to the free function namespace {TruncateRouteToEndAtSystem()}

This does not reference m_travel_route and does not modify the parameter
travel_route in place.

This passes in the owner instead of using this->Owner()

This does not tack on a system unknown to the empire to avoid returning
an empty path.

This fixes the following bugs:
1. In different locations the users of the function assumed that it used
m_travel_route as the source path or the parameter travel_route
2. In Fleet::Copy() the owner should be the new owner not the current
Owner()