Fix FleetWnd::Refresh location determination for selected ships.
As observed in review #1957 moving a fleet from a group of moving fleets
near each other but not at the exact same location, removes nearby
fleets from the FleetWnd.  This is because the check for all selected
ships in an exact location preceded the check for all ships in the same
vicinity.

This commit changes the order of location checks, so that the check for
selected ships in the same vicinity is earlier.