Fix MapWnd::FleetButtonLeftClicked over extending lifetime of FleetWnd
FleetWnd is a CUIWnd with its options stored in OptionsDB.
FleetWnd may also be a singular window depending on optons.
If FleetButtonLeftClicked creates a shared_ptr<FleetWnd> whose scope
extends past the subsequent SelectFleet(), then the FleetWnd can't be
deleted and recreated because the CUIWnd config will be unavailable.

This commit confines the lifetime of the shared_ptr<FleetWnd> to the if
clause that uses it.