Use the new AI fleet order OrderPause for exploration
When OrderMove was previously designated as the order type for exploration,
the AI fleet movement process forced them to stop at every system, and so a
simple MoveOrder was sufficient to cause an AI exploration fleet to visit,
and stop at, its target system.  When that move process was changed to no longer
always force stops at every system, the AI scouts would often be deemed to have
finished exploring a given system while actually still on their way to it,
and would get a new destination, causing them to merely pass through the
original target system without pausing.  This fixes that.