Fix focus change penalties
...so that they do not drop meter values below target.

Note that since the Stockpile currently uses a MaxStockpile rather than a
TargetStockpile, this entire penalty clause for Stockpile is currently
meaningless.  But since our past discussions of the Stockpile penalty for
focus changes had seemed to assume that the Stockpile used a TargetStockpile
meter, I am leaving this clause in place for now, until a clarifying
discussion can be had.