Fix the CombatReport log window initial scrollbar layout
Prior to this, when the CombatReport log window was first displayed,
the scrollbar button would not be visible/accessible even though the
full log could not fit within the displayed vertical space, nor would
mousescrolling work. Resizing the log window had been necessary to
enable scrolling.  With this fix, scrolling is enabled and the scrollbar
button is displayed from the outset.