Stop regeneration of Version.cpp in cases where it shouldn't
When the commit SHA could not be determined by the build scripts,
Version.cpp got regenerated whether an appropriate Version.cpp was
present or not.

This patch fixes that: now, when the commit SHA can't be
determined, but a Version.cpp is present, the version number is
checked against the one in the existing Version.cpp (instead of the
build number, which is the abbreviated commit SHA), and if the
version numbers match, Version.cpp is not regenerated.