Revert "Fix determinations of stale object visibilities"
This reverts commit 021786fddce87a2785048717840a9ddb1c764e8a.

PR #2147 fixed an issue with missing sensor ghosts by moving the stale object determination from an incremental adjustment per turn approach (which had not always been removing things from the stale list when glimpsed pre-combat or mid-combat) to a de novo determination per turn approach, but it turns out (in hindsight, it seems this should have been anticipated) the de novo approach inherently does not address situations where the object should still be known to be stale partly because of the conditions of a previous turn.

So, in preparation for following the alternate approach discussed in #2147, which was to have more than one incremental stale-object-update during a turn, this reverts the determination back to an incremental determination.