Describe NOT AND (A B..) as OR (NOT A NOT B..) by relying on negated description of A, B.. (inverse for NOR OR)
* Describe NOT AND (A B..) as  OR (NOT A NOT B..) by relying on negated description of A, B..
* Describe NOT  OR (A B..) as AND (NOT A NOT B..) by relying on negated description of A, B..
* there are stringtable entries for the NOT_AND, NOT_OR as maybe we want to be able to see a difference between a NOT AND and a OR.
  Also it could be that there is a wording difference in languages different from english. Else this could be removed.

Alternative implementation of DeMorgan via stringtables can be found in previous commit a4b44327b780d084d7721dc9abf3f6e0f38fc9a5

Reasoning:
I found its worthwhile to actually push the negation through to the operands because that gives the possibility for more helpful messages as double negation has a chance to cancel itself out (instead of '(neither an object that is not a ORB_SY...' you get a '(an object that is a ORB_SY and ..)'. Also in english this is really fine.