pre-warp era: make Remote Colonization researchable
Re-enable researching of the CON_REMOTE_COL technology.
Let it depend on GRO_PLANET_ECOL
to encourage research together with organic hulls.
Raise the research cost to the minimum of 8 RP;
no other technology is cheaper than that.
Update the syntax of its FOCS file to current standards.

Let GRO_LIFECYCLE_MAN depend on it because that unlocks
the subsequent upgrade to the colony pod.