pre-warp era: balancing tweaks to robotic hulls
Reduce the research cost and time of the Spatial Flux Bubble and
do not require an Orbital Drydock to build it, to allow meaningful
early choice in ship research.  Besides, it's a dead end anyway,
so making it cheap seems logical.

Let SHP_MIL_ROBO_CONT depend on SHP_BASIC_HULL_3 because
the Robotic Hull is excessively strong and cheap.