Improve AI estimation of standard enemy if no intel available.
The 6 damage comes from a MD4 which is a standard weapon found in early ships. This allows the AI to assess the value of early shields better. Otherwise, if it does not encounter any enemies for a long times, it will overvalue shields significantly, increasing costs whereas the later meeting of opponents usually means (basic) shields are worth less due to more time for improved weapon technology.