Introduce new AI MissionType: PROTECT_REGION
This new mission type is intended to not only target and defend a single system but rather its neighbors (and potentially more outwards systems) as well. The single system target (the primary objective of the mission) is, however, given priority.

If the primary objective is not immediately threatened, the protecting tries to engage and intercept nearby enemy fleets.