Use OrderedAlternativesOf condition for combat target tiers in fighter hangars
Notes:
This follows a public vote at  https://www.freeorion.org/forum/viewtopic.php?f=6&t=11204&p=94826#p94826

Filters targets with tiered choices (so only if the first tier has no matches, the next one is consulted).

The negation of OrderedAlternativesOf is quite complex, hopefully "Not OrderedAlternativesOf" has that connotation.

Interceptor targets are determined by the count of own Interceptor fighters in that battle multiplied with a spread factor of 1.2. This means if you have 10 interceptors in battle, all enemy fighters are sorted by their attack value and the topmost 12 enemy fighters will be used for the random choice of a target. This means that interceptors will probably will waste few shots and probably shoot down quite a number of the most dangerous enemies.

Bombers and Heavy Bomber fighters will shoot at enemy ships. Only if there are no enemy ships left at the beginning of a combat round, they will shoot at enemy fighters.

== GetDefaultInitialCandidateObjects ==

`GetDefaultInitialCandidateObjects` narrows the search space of objects when no list of candidates is provided.

If a single object or a containers of matches and non-matches are passed to `Eval`, then just those objects are tested by the condition. But there is an `Eval` overload that just takes a set of matches, which it immediately clears, and then fills with any objects in the universe that the condition matches.

To make this faster, conditions have the `GetDefaultInitialCandidateObjects` function to initialize the search space of objects to test, rather than always testing all objects in the `Universe`. For example, `Source` initializes with just the `ScriptingContext`'s source object, and `Type` will initialize based on what object type is being matched (eg. ships for `OBJ_SHIP`).

For this case, it's acting like an `Or` condition, and thus shouldn't narrow the search space at all, so like `Or` you shouldn't overload `GetDefaultInitialCandidateObjects` and instead just use the base class behaviour.

== SetTopLevelContent ==

`SetTopLevelContent` is used to set `m_top_level_content`, which is declared here:

https://github.com/freeorion/freeorion/blob/master/universe/ValueRef.h#L207

and which is used here:

https://github.com/freeorion/freeorion/blob/master/universe/ValueRef.cpp#L552

to return the top level bit of content in which this ValueRef is scripted. For example, in a species script, it would be the name of the species, returned as a `ValueRef::Constant<std::string>` when `ThisSpecies` is written into the script. This is helpful for writing generic macros that can be used in different contexts.

All the Effects, Conditions, and ValueRefs all have `SetTopLevelContent` functions, as they can contain eachother recursively and the final leaf ValueRef needs to have its top level content set by its enclosing object.