Try to merge fleets before generating fleet orders.
The idea is that the generated fleet orders may depend on the strength of a fleet. A military fleet may choose to engage the enemy only if it is stronger but flee from an enemy which is too powerful. If fleets are not merged before generating the orders, then the actual fleet strength available is not reflected correctly - all N fleets would be considered too weak to fight the enemy whereas the combined fleet is powerful enough for the fight and would thus generate a different (and seemingly more correct) set of orders.