Refactor: Use single temp_empty set in OrderedAlternativesOf MATCHES domain
With this refactoring there is no need to use matches in a hackish way anymore.
Having one temp var less means there is one extra move operation though.
That move is the happy path anyway (so just a single time happening) and it is based on moving, so i think the cleaner implementation is worth it.