Fix infinite loop in AI code
Triggered by an AI fleet that has a PROTECT_REGION mission and wants to keep bombarding a planet in a neighboring system.
When switching to a MILITARY mission, the fleet order mission may be canceled under certain conditions.
The fleet was then reassigned the PROTECT_REGION mission which is reassigning to a MILITARY mission which will be canceled again.

The fix is to just stay in the PROTECT_REGION mission and let the fleet stay in the current system.