Show SitRep for unlocked items when available (bit ugly)
As technology research was shifted after turn increase, SitReps for items unlocked by tech
are shown one turn after they are actually available (e.g. one turn after you can actually
enqueue Automated History Analyser).

This implementation fixes it by shifting the target turn for the unlock SitReps by minus one.
This makes a problem with the messages for initially unlocked items because those are registered
in minimal integer turn (e.g. -32767) which is 1+BEFORE_FIRST_TURN.

In order to fix this for turn zero or smaller I set the sitrep to a specific value.
If I set it to -1 , there are actually shown two different initial turns in the SitRep (one for
the real initial turn and one for -1)

This implementation fixes it by setting the sitrep turn to BEFORE_FIRST_TURN + 1

But usually for all that SitRep code, CurrentTurn()+1 is used so in future there might be a code
path for which the used CurrentTurn()+0 is wrong.

Also manipulating the turn number seems hackish.

So a better implementation could be to reverse the timing of the unlock by shifting it from
ApplyNewTechs to AddNewlyResearchedTechToGrantAtStartOfNextTurn.