Show descriptions for configured tags in the ship hull description
* shows description for tags on ship hulls
* which tags are described and in which order is defined in default/content_specific_parameters.txt
* the tag description is shown hardcoded below the hull description
* works well enough

TODO: Should rather add in VarText a tag for [[tags FUNCTIONAL_HULL_DESC_TAGS_LIST]]
which takes the FUNCTIONAL_HULL_DESC_TAGS_LIST, extracts HULL_DESC from the name, cycles through the actual tags of the hull
and checks if the <TAGNAME> is in the list. If so, look up the <HULL_DESC>_<TAGNAME> in the stringtables.
And revert the UI/EncyclopediaDetailPanel.cpp + str(FlexibleFormat(UserString("HULL_DESC")) + .. change

That would make the position of the list of tags configurable by the translator instead of the hardcoded position.