[WIP] Make OrderedAlternativesOf toplevel for Mass Driver combatTargets.
Currently only MD for testing (Oberlus playtest feedback sais the PR is not working yet)

I spotted the usual mistake of not making OrderedAlternativesOf toplevel. This could lead to
targeting untargetable targets in the first tier in the OrderedAlternativesOf (e.g. your own ship),
so not trying the second tier (so not including the defenseless planets).
COMBAT_TARGETS_VISIBLE_ENEMY would then remove all that targets which are not enemy (e.g. your ship) or not visible.