Detection Strength for Trith/Xenophobic Self Sustaining
The backend does not count heavily stealthed planets towards the
xenophia/self sustaining population malus, resulting in "unknown" status
when planets are beyond detection. Bug #2777

As it is not possible to determine the empire detection strength directly,
it is calculated from the available technologies, and then compared against
planets within the two jump distance to determine correct pop malus.

Signed-off-by: Rob Gill <rrobgill@protonmail.com>