Add new GG::Measure class to enable UI scaling
From the header:  This class is intended to encapsulate hard-coded
screen measurements to enable scaling.  A measure is simply a pixel
count, presumed to be a measure of distance along an axis for the
unscaled display.  The measure is scaled when converted to a GG::X or
GG::Y.

The scaling cannot be done during static initialisation because the
font size must be read from the options database.  The options
database is only populated when main() is run, after static
initialisation.  Therefore, we must store the unscaled raw value
(often provided in static constant declarations) and then scale the
value later when the GG::X or GG::Y is actually needed.