Fix bug in Layout::DoLayout calculating row heights and column widths
The issue is in the nesting of "if" statements while setting row and
columns widths.  The case of only a single row or column was
previously only treated specially within one of the branches of an
inner "if" statement when it needed to be treated in the outer "if"
statement.