Use expr instead of primary_expr when parsing NamedReal in the middle of another valueref - seems to fix precedence
Not sure what exactly is the difference, but it seems to fix the precedence and not to break anything.

Intended to fix issue #3188
  NamedReal/NamedInteger value declaration precedence is odd

Issue:

When writing this:
(NamedReal name = "GARRISON_2_MAXTROOPS_EFFECT" value = 2 * [[TROOPS_PER_POP]])

This is intended:
(NamedReal name = "GARRISON_2_MAXTROOPS_EFFECT" value = (2 * [[TROOPS_PER_POP]]))

But this is what we get:
((NamedReal name = "GARRISON_2_MAXTROOPS_EFFECT" value = 2) * [[TROOPS_PER_POP]])

Analysis was:

It seems precedence is working for integer, but not for floats:

(NamedReal name = "TWO_BY_THREE" value = 2 * 3) registers 2.0

(NamedReal name = "TWO_BY_THREE" value = 2.0 * 3.0) registers 2.0

(NamedInteger name = "TWO_BY_THREE" value = 2 * 3) registers 6

Also it seems this precedence is working for top level named value refs, so the only thing known to be broken is name-in-the-middle Real value refs.