Fix zero division in shield factor calculation see #3257
It is a rare issue that was reported by user. I have extracted logic to a separate function and add a protection for case when number of shots is zero.  It is not clear why does it happens, maybe  because of some species treats. Issue was introduced on turn 721, maybe it was some impure which was almost destroyed and its stats was not very actual.