Prefer linking to libOpenGL, fall back to libGL
CMP0072 has no practical effect here, is set to silence the warning.
It would effect the value of the OPENGL_LIBRARIES variable, but this one is not used.
The reason is likely that it additionally links to GLU and (with NEW) GLX libraries.

The used generator expression is available from cmake 3.8 on.