Use TEMPORARY_OBJECT_ID instead of unnamed magic number
EncyclopediaPanel uses InsertTemp (and turning off of signals) for similar purposes.
But that inserts a temporary object into the universe and at least at the moment
we get away by just using an id != -1 instead of adding and deleting it from the universe.
Potentially this could break when using subconditions(?)