FLANKING multi starlane attack bonus (aka pincer attack) via CountUnique PreviousToFinalDestinationID
Give flanking damage bonus to ships which were coming in from different starlanes on the previous turn
* effect is not given on direct jump
* Only works for end destinations (which is fine fluffwise, military needs to carefully plan the pincer attack)

=== Add PreviousToFinalDestinationID fleet property (also as valueref) ===

This returns the starlane/system a ship will arrive from when it reaches its destination.

Returns the second last system in the flight plan.
On the last jump it returns the previous system.
Returns invalid object if fleet is stationary.