Optimize AI systems_connected performance
Instead of running a BFS for each pair of system ids, cache the set of systems connected to the capital in a single depth-first-search and check if queried system_ids are within that set. Since most of the visible systems will be connected to the capital, this will cover most cases. Only if both systems are not connected to the capital, we will have to run another BFS. However, the search space will be significantly smaller since -again- most systems will be connected to the capital and thus won't be part of the search.