Remove reference in log to invariants in case of failing named ValueRef lookup
There was a reference to github issue (#3225 / indeterministic initialisation of UniverseObject invariants) when logging an error for failed named ValueRef lookup
"This should not happen if looking in the right registry once \"#3225 Refactor initialisation of invariants in value refs to happen after parsing\" is implemented"

Initialisation of invariants does not affect the registration (currently) - so even if it is done suboptimally/undeterministic (or if it gets changed in a later pass
in order to make it optimal/deterministic) its unrelated.
Only if we made registration of named values lazy (until the invariant is known) there might be connection.
So meanwhile we remove this.