[CI] Add merge to weekly release to the merge to master workflow (#4514)
I have created a separate workflow for merge to master and pull requests.
For the merge to master I have added one more step that is responsible for merging master to the weekly build and require manual approval.

This will allow releasing code by just finding the most latest build in the master and pressing the button. This step has one difference from the manual merge to the weekly build: you could do it only when all tests are passed. Because of test flakiness, I won't say that this is a big benefit. Good old manual merge and pus are still in place, so this should not be a problem.

As a result, we will have 3 levels of reusable workflows, this is a bit too much, but let's leave a bit with it and decide later if we should simplify it.